sqlite_to_postgres/requirements.txt
python-dotenv==0.21.0
psycopg2==2.9.4
